requirements.txt
# URC 2026 Critical Systems Dependencies

# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
psutil>=5.9.0
matplotlib>=3.7.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0

# Hardware mocking (for testing)
RPi.GPIO>=0.7.1
pyserial>=3.5
python-can>=4.3.1
spidev>=3.6

# ROS2 dependencies (when available)
rclpy>=3.3.0
geometry-msgs>=14.0.0
sensor-msgs>=4.3.0
nav-msgs>=1.1.0
std-msgs>=4.2.0

# Web/API dependencies (for mock hardware server)
flask>=3.0.0
requests>=2.31.0
aiofiles>=23.1.0
aiohttp>=3.8.0

# Data processing
pandas>=2.0.0
orjson>=3.9.0
structlog>=23.1.0
typer>=0.9.0

# Visualization and reporting
rich>=13.4.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# RealSense D435 driver (optional; for realsense_driver node)
pyrealsense2>=2.54.0

# Simulation dependencies (optional)
gazebo>=11.0.0
ros-humble-gazebo>=3.5.0
pybullet>=3.2.0

# Code quality
black>=22.10.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0